minor optimisation

+++ app/views/ITDC_Project/home/main/index.blade.php
@@ -6,32 +6,71 @@
         {{ Session::get('message') }}
     </div>
 @endif
-<table class="table table-bordered table-hover table-striped">
+<div class="fixedheader1">
+	<table class="table table-bordered table-hover table-striped">
+		<thead>
+			<th class="FLname">Firstname/Lastname</th>
+			<th class="gend1">Gender</th>
+			
+			<th class="skillshow">Skills</th>
+			@if(Auth::check())
+				<th class="action1">Action</th>
+			@endif
+		</thead>
+	</table>
+</div>
+<div class="scroll">
+	<table class="table table-bordered table-hover table-striped">
 		<tbody>
 			@foreach($users as $user)
 			<tr>
 				<td class="FLname">
-					<a href="{{ URL::to('admin/user/'.$user->id) }}">
+					<a href="{{ URL::route('user-profile', $user->username) }}">
 						{{ $user->username}}
 					</a>
 				</td>
 				<td class="gend1">{{ $user->gender }}</td>
-				<td class="edit1">
-					<a href="{{ URL::to('admin/user/'.$user->id.'/edit') }}" class="btn btn-primary btn-xs">
-						<i class="glyphicon glyphicon-pencil"></i>
-					</a>
+				<td class="skillshow">
+					@foreach($user->skills as $skill)
+						<?php $lvl = null; $seletced_skill = 'default'; $color_shade = null;  $s = null;
+							$lvl = $skill->pivot->level; $seletced_skill = 'info'; $s = ', LVL:';
+							$color_shade = 'style="background-color:hsl(120,40%,'.(100-$lvl).'%)"';
+						?>
+
+						<a href="{{ URL::to('#') }}" class="label label-{{ $seletced_skill }} bordered_1" {{ $color_shade }}>{{ $skill->name.$s.$lvl }}</a>
+					@endforeach
 				</td>
-				<td class="delete1">
-					{{ Form::open(array('route' => array('admin.user.destroy', $user->id), 'method' => 'delete', 'class' => 'delete_user')) }}
-			    		<button type="submit" class="btn btn-default btn-xs">
-			    			<i class="glyphicon glyphicon-remove text-danger"></i>
-			    		</button>
-					{{ Form::close() }}
-				</td>
+				@if(Auth::check())
+					<td class="edit1">
+						<a href="{{ URL::to('admin/user/'.$user->id.'/edit') }}" class="btn btn-primary btn-xs">
+							<i class="glyphicon glyphicon-pencil"></i>
+						</a>
+					</td>
+					<td class="delete1">
+						{{ Form::open(array('route' => array('admin.user.destroy', $user->id), 'method' => 'delete', 'class' => 'delete_user')) }}
+				    		<button type="submit" class="btn btn-default btn-xs">
+				    			<i class="glyphicon glyphicon-remove text-danger"></i>
+				    		</button>
+						{{ Form::close() }}
+					</td>
+				@endif
 			</tr>
 			@endforeach
 		</tbody>
 	</table>
+</div>
+
+
+	<div class="">
+		{{$users->appends(Input::all())->links()}}
+	</div>
+
+
+	<script type="text/javascript">
+
+	/////// pagination wrap
+		
+	</script>
 @stop
 
 <!-- ////////////////////////////////////////////////////// -->
